+++ releases.md
@@ -1,3 +1,5 @@
 # Releases
 
 ## Unreleased
+
+  - Dump screenshot and HTML of the browser session on test failure for easier debugging.